requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gspread==3.6.0
httplib2==0.17.3
idna==2.9
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==2.0
uritemplate==3.0.1
urllib3==1.25.9
